Refactor: result module
- 重构 result 为模块
