+++ omega_miya/result.py
@@ -1,41 +1,118 @@
 """
 @Author         : Ailitonia
 @Date           : 2022/02/20 17:04
-@FileName       : __init__.py.py
+@FileName       : result.py
 @Project        : nonebot2_miya 
 @Description    : Omega Results Models and Tools
 @GitHub         : https://github.com/Ailitonia
 @Software       : PyCharm 
 """
 
+
 import inspect
+from datetime import date
 from functools import wraps
-from typing import TypeVar, ParamSpec, Callable, Coroutine
+from typing import Dict, List, Set, Tuple, Any, Optional, TypeVar, Generic, ParamSpec, Callable, Coroutine
+from pydantic import Json
+from pydantic.generics import GenericModel
 from nonebot import logger
-from .results import (
-    BaseResult,
-    DictResult,
-    IntDictResult,
-    TextDictResult,
-    JsonDictResult,
-    ListResult,
-    IntListResult,
-    TextListResult,
-    TupleListResult,
-    DictListResult,
-    DateListResult,
-    SetResult,
-    IntSetResult,
-    TextSetResult,
-    TupleResult,
-    IntTupleResult,
-    TextTupleResult,
-    IntResult,
-    TextResult,
-    BoolResult,
-    BytesResult,
-    AnyResult
-)
+
+
+T = TypeVar('T')
+
+
+class BaseResult(GenericModel, Generic[T]):
+    error: bool
+    info: str
+    result: Optional[T] = None
+
+    @property
+    def success(self) -> bool:
+        return not self.error
+
+
+class DictResult(BaseResult):
+    result: dict
+
+
+class IntDictResult(BaseResult):
+    result: Dict[int, int]
+
+
+class TextDictResult(BaseResult):
+    result: Dict[str, str]
+
+
+class JsonDictResult(BaseResult):
+    result: Json
+
+
+class ListResult(BaseResult):
+    result: list
+
+
+class IntListResult(BaseResult):
+    result: List[int]
+
+
+class TextListResult(BaseResult):
+    result: List[str]
+
+
+class TupleListResult(BaseResult):
+    result: List[tuple]
+
+
+class DictListResult(BaseResult):
+    result: List[dict]
+
+
+class DateListResult(BaseResult):
+    result: List[date]
+
+
+class SetResult(BaseResult):
+    result: set
+
+
+class IntSetResult(BaseResult):
+    result: Set[int]
+
+
+class TextSetResult(BaseResult):
+    result: Set[str]
+
+
+class TupleResult(BaseResult):
+    result: tuple
+
+
+class IntTupleResult(BaseResult):
+    result: Tuple[int, ...]
+
+
+class TextTupleResult(BaseResult):
+    result: Tuple[str, ...]
+
+
+class IntResult(BaseResult):
+    result: int
+
+
+class TextResult(BaseResult):
+    result: str
+
+
+class BoolResult(BaseResult):
+    result: bool
+
+
+class BytesResult(BaseResult):
+    result: bytes
+
+
+class AnyResult(BaseResult):
+    result: Any
 
 
 P = ParamSpec("P")
